requirements2.txt
boto3==1.34.114
channels==4.0.0
cryptography==41.0.7
diskcache==5.6.3
Django==5.0.7
django_phonenumber_field==8.0.0
djangorestframework_simplejwt==5.3.1
duckdb==0.9.2
earthengine_api==0.1.384
GDAL==3.4.1
geemap==0.32.1
GeoAlchemy2==0.15.2
geopandas==0.14.4
google_api_python_client==2.116.0
google_auth_oauthlib==1.0.0
imageio==2.34.2
ipython==8.12.3
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.10
langchain_google_genai==2.0.1
langchain_openai==0.2.2
leafmap==0.30.0
llama_index==0.11.17
matplotlib==3.9.2
minio==7.2.5
more_itertools==10.5.0
numpy==1.23.5
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
phonepe_sdk==1.1.0
Pillow==10.4.0
protobuf==5.28.2
psycopg2==2.9.9
psycopg2_binary==2.9.9
pygdal==3.6.4.11
pygsheets==2.0.6
pyproj==3.7.0
pyshorteners==1.0.1
python_crontab==3.2.0
rasterio==1.3.8
Requests==2.32.3
requests_toolbelt==1.0.0
segment_geospatial==0.10.2
Shapely==2.0.6
SQLAlchemy==2.0.23
sympy==1.13.1
tifffile==2024.7.24
torch==2.4.0
turfpy==0.0.7
ultralytics==8.1.14
##
# langchain
# langchain_community
# langchain_core
# langchain_google_genai
# langchain_openai
# llama_index
# matplotlib
# more_itertools
# pandas
# Pillow
# protobuf
# requests
# Shapely
# up42
#These libraries has to be installed manually
##
